Allow overriding the "official" URI for routing.
Used if you want to mess with it in a subclass. This can be useful
if you want to, for example, transparently map multiple URIs to the
same route. (Note it is *generally* better to use redirects as
search engines ostensibly do not like multiple pages on a site with
identical content.)